refactor: Move void return type check to PolymerServerEventHandlers
Moved the void return type validation from AbstractServerHandlers to
PolymerServerEventHandlers where it's actually needed. The base class
now has an empty default implementation allowing all return types.
Removed redundant override in ClientCallableHandlers.
